+++ open-codegen/opengen/constraints/ball1.py
@@ -6,16 +6,18 @@ import numpy as np
 class Ball1(Constraint):
     """Ball1 aka Norm-1 Ball
 
+    Ball-1 with given radius.
+
     """
 
     def __init__(self, radius: float = 1.0):  # unless specified, alpha=1
         """Constructor for a Ball1
 
         Args:
-            radius: ball radius (default: 1)
+            :param radius: ball radius (default: 1)
 
         Returns:
-            New instance of Ball1 with given radius
+            :return: New instance of Ball1 with given radius
         """
         if radius <= 0:
             raise ValueError("radius must be a positive number")
@@ -31,6 +33,15 @@ class Ball1(Constraint):
         raise NotImplementedError()
 
     def project(self, u):
+        """Project on the current Ball-1
+
+        Args:
+            :param u: vector u
+
+        Returns:
+            :return: projection of u onto the current ball-1
+
+        """
         if np.linalg.norm(u, 1) <= self.radius:
             return u